// pages/api/users/books/books-model.ts
export interface Book {
    id: number;
    title: string;
    description: string;
    image_url: string;
    release_year: number;
    price: string;
    total_page: number;
    thickness: string;
    created_at: Date;
    updated_at: Date;
    category_id: number;
  }
  
  export type BookInput = Omit<Book, 'id' | 'created_at' | 'updated_at'>;